callingninja-api-fastapi/requirements.txt
aioboto3==11.2.0
aiobotocore==2.5.0
aiohttp==3.8.4
aiohttp-retry==2.8.3
aioitertools==0.11.0
aiosignal==1.3.1
anyio==3.7.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
boto3==1.26.76
botocore==1.29.76
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
decorator==5.1.1
dnspython==2.3.0
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.97.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
ipython==8.14.0
jedi==0.18.2
jmespath==1.0.1
matplotlib-inline==0.1.6
motor==3.1.2
multidict==6.0.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.9
Pygments==2.15.1
PyJWT==2.7.0
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.27.0
traitlets==5.9.0
twilio==8.2.2
typing_extensions==4.6.3
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
wrapt==1.15.0
yarl==1.9.2
